+++ deep_conmech/graph/data/data_scenario.py
@@ -1,8 +1,10 @@
 from deep_conmech.graph.data.data_base import *
-from deep_conmech.common import config, trh
+from deep_conmech.common import config
 from deep_conmech.simulator.calculator import Calculator
 from deep_conmech.simulator.setting.setting_forces import *
 from deep_conmech.graph.setting.setting_input import SettingInput
+from deep_conmech.graph.helpers import thh
+from conmech.helpers.mph import mph
 
 
 def get_setting(scenario):
@@ -95,7 +97,7 @@ class TrainingScenariosDatasetDynamic(BaseDatasetDynamic):
 
         result = False
         while result is False:
-            result = thh.run_processes(
+            result = mph.run_processes(
                 generate_scenario_data_process, (self, self.all_scenarios), num_workers,
             )
             if result is False:
@@ -115,6 +117,6 @@ class ValidationDatasetDynamic(BaseDatasetDynamic):
     def generate_all_data(self):
         result = False
         while result is False:
-            result = thh.run_processes(
+            result = mph.run_processes(
                 generate_scenario_data_process, (self, [self.scenario]), 1,
             )
