+++ deep_conmech/graph/helpers/thh.py
@@ -3,7 +3,6 @@ import resource
 import time
 from ctypes import ArgumentError
 from datetime import datetime
-from multiprocessing import Process, Queue
 
 import numpy as np
 import pandas
@@ -157,40 +156,3 @@ def create_folders(path):
         create_folder(final_path)
 
 
-
-
-def run_processes(function, function_args, num_workers):
-    queue = Queue()
-
-    processes = [
-        Process(target=lambda *args: queue.put(function(*args)), args=function_args + (num_workers, process_id))
-        for process_id in range(num_workers)
-    ]
-
-    for p in processes:
-        p.start()
-
-    for p in processes:
-        p.join()
-
-    for _ in processes:
-        done = queue.get()
-        if not done:
-            return False
-
-    return True
-
-
-
-def run_process(function): #, args
-    queue = Queue()
-
-    #wrapper = lambda *args : queue.put(function())
-    wrapper = lambda: queue.put(function())
-
-    process = Process(target=wrapper)
-    process.start()
-    process.join()
-    result = queue.get()
-
-    return result
